Switching Duration to Float since that is the new Datatype

+++ src/org/vitrivr/cineast/core/db/MultimediaObjectLookup.java
@@ -70,7 +70,7 @@ public class MultimediaObjectLookup{
 			return new MultimediaObjectDescriptor();
 		}
 		
-		if(!checkProvider("duration", durationProvider, ProviderDataType.DOUBLE)){
+		if(!checkProvider("duration", durationProvider, ProviderDataType.FLOAT)){
 			return new MultimediaObjectDescriptor();
 		}
